requirements.txt
absl-py==1.3.0
aio-pika==8.2.3
aiofiles==24.1.0
aiogram==2.25.2
aiohttp==3.8.6
aiormq==6.4.2
aiosignal==1.3.1
airportsdata==20240415
airtable-python-wrapper==0.15.3
amadeus==10.1.0
aniso8601==7.0.0
annoying==0.1.0
anyascii==0.3.2
anyio==4.4.0
APScheduler==3.9.1
argh==0.31.3
asgiref==3.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.1.0
Babel==2.9.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
boto3==1.34.150
Brotli==1.1.0
build==1.2.1
CacheControl==0.12.14
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
confusable-homoglyphs==3.1.1
cryptography==43.0.1
cssselect2==0.7.0
cycler==0.12.1
dask==2022.10.2
defusedxml==0.8.0rc2
diff-match-patch==20230430
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.7
django-activity-stream==2.0.0
django-allauth==64.0.0
django-annoying==0.10.6
django-baton==4.2.2
django-bootstrap3==24.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.13
django-cors-headers==4.4.0
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-dirtyfields==1.9.3
django-export==1.11.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.1.1
django-js-asset==2.2.0
django-livereload-server==0.5.1
django-manager-utils==3.1.5
django-modelcluster==6.3
django-object-tools==2.0.0
django-permissionedforms==0.1
django-query-builder==3.2.0
django-registration==3.1.2
django-registration-redux==2.13
django-rosetta==0.10.0
django-taggit==5.0.1
django-treebeard==4.7.1
django_csp==3.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==1.16.0
docker-py==1.9.0
docopt==0.6.2
docutils==0.21.2
draftjs-exporter==5.0.0
drf-yasg==1.21.8
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
factory_boy==3.3.1
fbmessenger==6.0.0
filetype==1.2.0
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.32.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==3.0.3
grpcio==1.65.1
gunicorn==19.9.0
h11==0.14.0
h5py==3.11.0
hgicommon==1.2.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==3.7
import-export==0.3.1
inflection==0.5.1
isodate==0.6.1
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.4
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.5
l18n==2021.3
laces==0.1.1
libclang==18.1.1
livereload==2.7.0
locket==1.0.0
magic-filter==1.0.12
Markdown==3.6
MarkupSafe==2.1.5
mattermostwrapper==2.2
mccabe==0.7.0
msgpack==1.0.8
multidict==5.2.0
networkx==2.6.3
nltk==3.8.1
numpy==2.2.0
oauthlib==3.2.2
openai==1.37.0
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
pamqp==3.2.1
pandas==2.2.2
paramiko==2.8.1
partd==1.4.2
pillow==10.4.0
pillow_heif==0.18.0
pip-tools==7.4.1
pip3-autoremove==1.2.2
plotly==5.24.1
pluggy==1.5.0
polib==1.2.0
portalocker==2.10.1
promise==2.3
prompt-toolkit==2.0.9
protobuf==3.19.6
psutil==6.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycountry==24.6.1
pycparser==2.22
pydantic==1.9.0
pydot==1.4.2
pydyf==0.11.0
PyJWT==2.8.0
pykwalify==1.8.0
pymongo==3.10.1
PyNaCl==1.5.0
pyparsing==3.1.2
PyPDF2==1.28.6
pyphen==0.16.0
pyproject_hooks==1.1.0
pyreadline3==3.4.1
pyrsistent==0.20.0
python-crfsuite==0.9.10
python-decouple==3.8
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0.2
questionary==1.10.0
randomname==0.1.5
redis==4.5.3
regex==2022.10.31
reportlab==4.2.5
requests==2.31.0
requests-oauthlib==1.2.0
requests-toolbelt==1.0.0
rocketchat-API==1.28.1
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
Rx==1.6.3
s3transfer==0.10.2
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.14.1
sentry-sdk==1.14.0
simple-websocket==1.0.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack_sdk==3.31.0
sniffio==1.3.1
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.1
sshtunnel==0.4.0
tablib==3.5.0
tabulate==0.9.0
tarsafe==0.0.3
telepath==0.3.1
tenacity==9.0.0
termcolor==2.4.0
terminaltables==3.1.10
textblob==0.18.0.post0
threadpoolctl==3.5.0
tinycss2==1.3.0
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
tweepy==4.14.0
twilio==7.14.2
typeguard==4.3.0
typing-utils==0.1.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
Unipath==1.1
uritemplate==4.1.1
urllib3==2.2.2
uvloop==0.20.0
wagtail==6.2
watchdog==0.8.3
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
webexteamssdk==1.6.1
websocket-client==1.8.0
websockets==10.4
whitenoise==6.7.0
Willow==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zopfli==0.2.3
